[#14] feat: 도넛 그래프 컴포넌트 분리
