+++ src/components/summoner/match/matchSummary/SummaryInfo.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import styled from "styled-components";
 import { Doughnut } from "react-chartjs-2";
 
+import DoughnutGraph from "./DoughnutGraph";
+
 const SummaryInfoWrap = styled.div`
   padding: 16px 24px;
   border-right: 1px solid #cdd2d2;
@@ -13,7 +15,7 @@ const WinLossRatioWrap = styled.div`
   text-align: center;
 `;
 
-const DoughnutWrap = styled.div`
+const GraphWrap = styled.div`
   margin-top: 14px;
   width: 90px;
   height: 90px;
@@ -23,41 +25,13 @@ const SummaryInfo = ({ summary }) => {
   const { wins, losses, kills, deaths, assists } = summary;
   const games = wins + losses;
 
-  const options = {
-    cutoutPercentage: 70,
-    maintainAspectRatio: false,
-    legend: {
-      display: false,
-    },
-    elements: {
-      arc: {
-        borderWidth: 0,
-      },
-    },
-    tooltips: {
-      enabled: false,
-    },
-  };
-
-  const colors = ["#ee5a52", "#1f8ecd"];
-
-  const data = {
-    datasets: [
-      {
-        data: [losses, wins],
-        backgroundColor: colors,
-        hoverBackgroundColor: colors,
-      },
-    ],
-  };
-
   return (
     <SummaryInfoWrap>
       <WinLossRatioWrap>
         {games}전 {wins}승 {losses}패
-        <DoughnutWrap>
-          <Doughnut data={data} options={options} />
-        </DoughnutWrap>
+        <GraphWrap>
+          <DoughnutGraph {...{ wins, losses }} />
+        </GraphWrap>
       </WinLossRatioWrap>
     </SummaryInfoWrap>
   );
